Basic
Added the base for the premium accounts module. The idea is to give players certain rewards, enabled using bitmask. It could, for instance, be used to manage donation rewards tied to an account.

There's still a lot of work left to do as I've only added three example items to help me test the module as I'm working on it.

I must advise against using the module until it's more complete as I'll be adding items and possibly move the bits around.

I also have to plan the message sent with the item in the mail.